+++ apps/common-app/src/apps/reanimated/examples/index.ts
@@ -3,6 +3,9 @@
 import 'react-native-reanimated';
 import React from 'react';
 
+import type { Example } from '@/components';
+import { REAPlatform } from '@/components';
+
 const AboutExample: React.FC = () =>
   React.createElement(require('./AboutExample').default as React.FC);
 const AmountExample: React.FC = () =>
@@ -117,6 +120,8 @@ const ColorInterpolationExample: React.FC = () =>
   React.createElement(
     require('./ColorInterpolationExample').default as React.FC
   );
+const ContrastColorExample: React.FC = () =>
+  React.createElement(require('./ContrastColorExample').default as React.FC);
 const CombinedTest: React.FC = () =>
   React.createElement(
     require('./LayoutAnimations/Combined').default as React.FC
@@ -132,10 +137,6 @@ const ComposedHandlerDifferentEventsExample: React.FC = () =>
 const ComposedHandlerInternalMergingExample: React.FC = () =>
   React.createElement(
     require('./ComposedHandlerInternalMergingExample').default
-  );
-const CopySerializablePerformanceTest: React.FC = () =>
-  React.createElement(
-    require('./CopySerializablePerformanceTest').default as React.FC
   );
 const CounterExample: React.FC = () =>
   React.createElement(require('./CounterExample').default as React.FC);
@@ -202,10 +203,12 @@ const NestedExitingCleanupExample: React.FC = () =>
   );
 const ExtrapolationExample: React.FC = () =>
   React.createElement(require('./ExtrapolationExample').default as React.FC);
-const FetchExample: React.FC = () =>
-  React.createElement(require('./FetchExample').default as React.FC);
 const FilterExample: React.FC = () =>
   React.createElement(require('./FilterExample').default as React.FC);
+const FinalFrameAccuracyExample: React.FC = () =>
+  React.createElement(
+    require('./LayoutAnimations/FinalFrameAccuracy').default as React.FC
+  );
 const FlatListExample: React.FC = () =>
   React.createElement(
     require('./SharedElementTransitions/FlatList').default as React.FC
@@ -238,10 +241,6 @@ const HabitsExample: React.FC = () =>
   React.createElement(
     require('./LayoutAnimations/HabitsExample').default as React.FC
   );
-const HermesSamplingProfilerExample: React.FC = () =>
-  React.createElement(
-    require('./HermesSamplingProfilerExample').default as React.FC
-  );
 const IPodExample: React.FC = () =>
   React.createElement(require('./IPodExample').default as React.FC);
 const ImageStackExample: React.FC = () =>
@@ -270,6 +269,11 @@ const LayoutAnimationExample: React.FC = () =>
   React.createElement(
     require('./SharedElementTransitions/LayoutAnimation').default
   );
+const LayoutAnimationBatchSyncExample: React.FC = () =>
+  React.createElement(
+    require('./LayoutAnimations/LayoutAnimationBatchSyncExample')
+      .default as React.FC
+  );
 const LayoutTransitionExample: React.FC = () =>
   React.createElement(
     require('./LayoutAnimations/LayoutTransitionExample').default
@@ -284,8 +288,6 @@ const ListItemLayoutAnimation: React.FC = () =>
   React.createElement(
     require('./LayoutAnimations/ListItemLayoutAnimation').default
   );
-const LogExample: React.FC = () =>
-  React.createElement(require('./LogExample').default as React.FC);
 const ManyScreensExample: React.FC = () =>
   React.createElement(
     require('./SharedElementTransitions/ManyScreens').default
@@ -312,8 +314,6 @@ const ModalsExample: React.FC = () =>
   React.createElement(
     require('./SharedElementTransitions/Modals').default as React.FC
   );
-const ModifyExample: React.FC = () =>
-  React.createElement(require('./ModifyExample').default as React.FC);
 const MountingUnmounting: React.FC = () =>
   React.createElement(
     require('./LayoutAnimations/MountingUnmounting').default as React.FC
@@ -466,12 +466,8 @@ const SwipeableListExample: React.FC = () =>
   React.createElement(require('./SwipeableListExample').default as React.FC);
 const SyncBackToReactExample: React.FC = () =>
   React.createElement(require('./SyncBackToReactExample').default as React.FC);
-const SynchronizablePerformanceExample: React.FC = () =>
-  React.createElement(require('./SynchronizableExample').default as React.FC);
 const SynchronousPropsExample: React.FC = () =>
   React.createElement(require('./SynchronousPropsExample').default as React.FC);
-const SystraceSectionExample: React.FC = () =>
-  React.createElement(require('./SystraceSectionExample').default as React.FC);
 const TabNavigatorExample: React.FC = () =>
   React.createElement(
     require('./SharedElementTransitions/TabNavigatorExample').default
@@ -508,25 +504,6 @@ const WithoutBabelPluginExample: React.FC = () =>
   );
 const WobbleExample: React.FC = () =>
   React.createElement(require('./WobbleExample').default as React.FC);
-
-export const REAPlatform = {
-  IOS: 'ios',
-  ANDROID: 'android',
-  MACOS: 'macos',
-  WEB: 'web',
-};
-
-export interface Example {
-  icon?: string;
-  title: string;
-  screen: React.FC;
-  shouldWork?: {
-    ios: boolean;
-    android: boolean;
-  };
-  disabledPlatforms?: (typeof REAPlatform)[keyof typeof REAPlatform][];
-  needsBundleMode?: boolean;
-}
 
 export const EXAMPLES: Record<string, Example> = {
   // About
@@ -568,16 +545,6 @@ export const EXAMPLES: Record<string, Example> = {
     title: 'Slow animations',
     screen: SlowAnimationsExample,
   },
-  HermesSamplingProfilerExample: {
-    icon: '📊',
-    title: 'Hermes sampling profiler',
-    screen: HermesSamplingProfilerExample,
-  },
-  SystraceSectionExample: {
-    icon: '📊',
-    title: 'Systrace section',
-    screen: SystraceSectionExample,
-  },
   SyncBackToReactExample: {
     icon: '🔄',
     title: 'Sync back to React',
@@ -603,27 +570,10 @@ export const EXAMPLES: Record<string, Example> = {
     title: 'Third party components',
     screen: ThirdPartyComponentsExample,
   },
-  Synchronizable: {
-    icon: '🔄',
-    title: 'Synchronizable performance',
-    screen: SynchronizablePerformanceExample,
-    disabledPlatforms: [REAPlatform.WEB],
-  },
   ReactFreeze: {
     icon: '❄️',
     title: 'React freeze',
     screen: FreezeExample,
-  },
-  FetchExample: {
-    icon: '📡',
-    title: 'Fetch & XHR (Bundle Mode)',
-    screen: FetchExample,
-    needsBundleMode: true,
-  },
-  ModifyExample: {
-    icon: '🪛',
-    title: 'Modify',
-    screen: ModifyExample,
   },
   CircularSliderExample: {
     icon: '🔘',
@@ -649,12 +599,6 @@ export const EXAMPLES: Record<string, Example> = {
     icon: '⬆️',
     title: 'Bottom sheet',
     screen: BottomSheetExample,
-  },
-  CopySerializablePerformanceTest: {
-    icon: '🔄',
-    title: 'Copy serializable performance test',
-    screen: CopySerializablePerformanceTest,
-    disabledPlatforms: [REAPlatform.WEB],
   },
   FlatListWithLayoutAnimations: {
     icon: '🎻',
@@ -963,11 +907,6 @@ export const EXAMPLES: Record<string, Example> = {
       REAPlatform.MACOS,
     ],
   },
-  LogExample: {
-    icon: '⌨',
-    title: 'Log test',
-    screen: LogExample,
-  },
   HabitsExample: {
     icon: '🧑‍💻',
     title: 'Habits',
@@ -1073,6 +1012,11 @@ export const EXAMPLES: Record<string, Example> = {
     title: 'Color interpolation',
     screen: ColorInterpolationExample,
   },
+  ContrastColorExample: {
+    icon: '🔲',
+    title: 'Contrast color',
+    screen: ContrastColorExample,
+  },
   ExtrapolationExample: {
     title: 'Extrapolation example',
     screen: ExtrapolationExample,
@@ -1122,6 +1066,10 @@ export const EXAMPLES: Record<string, Example> = {
   BasicLayoutAnimation: {
     title: '[LA] Basic layout animation',
     screen: BasicLayoutAnimation,
+  },
+  LayoutAnimationBatchSync: {
+    title: '[LA] Batch synchronization',
+    screen: LayoutAnimationBatchSyncExample,
   },
   BasicNestedAnimation: {
     title: '[LA] Basic nested animation',
@@ -1263,6 +1211,10 @@ export const EXAMPLES: Record<string, Example> = {
     icon: '🧹',
     title: '[LA] Nested exiting cleanup',
     screen: NestedExitingCleanupExample,
+  },
+  FinalFrameAccuracyExample: {
+    screen: FinalFrameAccuracyExample,
+    title: '[LA] Final frame accuracy',
   },
 
   // Shared Element Transitions