+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/common/values/CSSValueVariant.cpp
@@ -99,7 +99,7 @@ template <CSSValueDerived... AllowedTypes>
 CSSValueVariant<AllowedTypes...> CSSValueVariant<AllowedTypes...>::interpolate(
     const double progress,
     const CSSValueVariant &to,
-    const ValueInterpolationContext &context) const {
+    const InterpolationContextFor<AllowedTypes...> &context) const {
   if (storage_.index() != to.storage_.index()) {
     return fallbackInterpolate(progress, to, context.fallbackInterpolateThreshold);
   }
@@ -107,34 +107,12 @@ CSSValueVariant<AllowedTypes...> CSSValueVariant<AllowedTypes...>::interpolate(
   return std::visit(
       [&](const auto &fromValue, const auto &toValue) -> CSSValueVariant {
         REA_IF_SAME_TYPE(fromValue, toValue) {
-          if constexpr (Resolvable<L>) {
-            throw std::runtime_error("[Reanimated] Resolvable value cannot be interpolated as non-resolvable");
-          } else if (fromValue.canInterpolateTo(toValue)) {
-            return CSSValueVariant(fromValue.interpolate(progress, toValue));
-          }
-        }
-        return fallbackInterpolate(progress, to, context.fallbackInterpolateThreshold);
-      },
-      storage_,
-      to.storage_);
-}
-
-template <CSSValueDerived... AllowedTypes>
-CSSValueVariant<AllowedTypes...> CSSValueVariant<AllowedTypes...>::interpolate(
-    const double progress,
-    const CSSValueVariant &to,
-    const ResolvableValueInterpolationContext &context) const {
-  if (storage_.index() != to.storage_.index()) {
-    return fallbackInterpolate(progress, to, context.fallbackInterpolateThreshold);
-  }
-
-  return std::visit(
-      [&](const auto &fromValue, const auto &toValue) -> CSSValueVariant {
-        REA_IF_SAME_TYPE(fromValue, toValue) {
-          if constexpr (!Resolvable<L>) {
-            throw std::runtime_error("[Reanimated] Non-resolvable value cannot be interpolated as resolvable");
-          } else if (fromValue.canInterpolateTo(toValue)) {
-            return CSSValueVariant(fromValue.interpolate(progress, toValue, context));
+          if (fromValue.canInterpolateTo(toValue)) {
+            if constexpr (requires { fromValue.interpolate(progress, toValue, context); }) {
+              return CSSValueVariant(fromValue.interpolate(progress, toValue, context));
+            } else {
+              return CSSValueVariant(fromValue.interpolate(progress, toValue));
+            }
           }
         }
         return fallbackInterpolate(progress, to, context.fallbackInterpolateThreshold);
@@ -155,6 +133,7 @@ template class CSSValueVariant<CSSLength>;
 template class CSSValueVariant<CSSLength, CSSKeyword>;
 template class CSSValueVariant<CSSDouble>;
 template class CSSValueVariant<CSSDouble, CSSKeyword>;
+template class CSSValueVariant<CSSTextDouble>;
 template class CSSValueVariant<CSSInteger>;
 template class CSSValueVariant<CSSIndex>;
 template class CSSValueVariant<CSSKeyword>;