+++ packages/react-native-worklets/src/memory/synchronizable.native.ts
@@ -2,16 +2,50 @@
 
 import { WorkletsModule } from '../WorkletsModule/NativeWorklets';
 import { createSerializable } from './serializable';
-import type { Synchronizable } from './types';
+import type {
+  FixedSynchronizable,
+  Synchronizable,
+  SynchronizableConfig,
+} from './types';
+
+export function createSynchronizable<TValue extends number | boolean>(
+  initialValue: TValue,
+  config: SynchronizableConfig & { fixedType: true }
+): FixedSynchronizable<TValue extends boolean ? boolean : number>;
+
+export function createSynchronizable<TValue extends number | boolean>(
+  initialValue: TValue,
+  config: SynchronizableConfig
+): Synchronizable<TValue> | FixedSynchronizable<TValue>;
 
 export function createSynchronizable<TValue = unknown>(
-  initialValue: TValue
+  initialValue: TValue,
+  config?: SynchronizableConfig
+): Synchronizable<TValue>;
+
+export function createSynchronizable<TValue = unknown>(
+  initialValue: TValue,
+  config?: SynchronizableConfig
 ): Synchronizable<TValue> {
+  const isFixed = !!config?.fixedType;
+
+  if (
+    __DEV__ &&
+    isFixed &&
+    !(typeof initialValue === 'number' || typeof initialValue === 'boolean')
+  ) {
+    throw new Error(
+      '[Worklets] `fixedType` requires a number or boolean initial value.'
+    );
+  }
+
   const synchronizableRef = WorkletsModule.createSynchronizable(
-    createSerializable(initialValue)
+    isFixed ? initialValue : createSerializable(initialValue),
+    isFixed
   );
 
   return globalThis.__synchronizableUnpacker(
-    synchronizableRef
+    synchronizableRef,
+    isFixed
   ) as unknown as Synchronizable<TValue>;
 }